backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose
pymongo
